requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
async-timeout==4.0.2
-e git+https://github.com/Kozphy/bot.git@6bc2fa7af9ed97da98007ae56c87960bd18168cb#egg=attr_BOT_NAME
attrs==21.4.0
bbgo @ file:///home/kopher/go/src/github.com/c9s/bbgo/python
-e git+https://github.com/Kozphy/bot.git@6bc2fa7af9ed97da98007ae56c87960bd18168cb#egg=bot
build==0.7.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
flake8==4.0.1
frozenlist==1.3.0
greenlet==1.1.2
grpcio==1.46.0
grpcio-tools==1.46.0
idna==3.3
kucoin-python==1.0.11
loguru==0.6.0
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
mysqlclient==2.1.0
numpy==1.22.3
packaging==21.3
pep517==0.12.0
protobuf==3.20.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pyrate-limiter==2.8.1
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
SQLAlchemy-Utils==0.38.2
tomli==2.0.1
urllib3==1.26.9
websockets==10.3
yarl==1.7.2
